client-go/util/consistencydetector: update after moving to the new package
