+++ staging/src/k8s.io/client-go/util/consistencydetector/data_consistency_detector.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cache
+package consistencydetector
 
 import (
 	"context"
